requirements.txt
azure-storage-blob>=12.0.0
python-dotenv>=0.19.0
pandas>=1.3.0
matplotlib>=3.4.0
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
